requirements.txt
jsonpickle
pillow